+++ src/view/global/card/index.ts
