+++ src/view/global/footer/index.ts
