+++ src/view/global/link/index.ts
@@ -1,4 +1,4 @@
-import bloggerHref from '../../utils/blogger/href'
+import bloggerHref from '../../../utils/blogger/href'
 
 interface TArgs {
   name: string
