+++ src/view/global/nav/index.ts
@@ -1,4 +1,4 @@
-import type { TLinksNav } from '../../configs/links/nav'
+import type { TLinksNav } from '../../../configs/links/nav'
 import Link from '../link'
 import NavBottom from './bottom'
